pyproject.toml
[tool.poetry]
name = "nagasaki"
version = "0.1.0"
description = ""
authors = ["Oskar Korczak <oskar.korczak@protonmail.com>", "Kacper Wikieł <k.wikiel@gmail.com>", "Szczepan Adriański <kontakt@aszczepanski.pl>"]
 

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.2.5"
pydantic = "^1.8.2"
python-dotenv = "^0.19.1"
responses = "^0.15.0"
websocket-client = "^1.2.1"
pandas = "^1.3.4"
APScheduler = "^3.8.1"
pylint = "^2.12.2"
hypothesis = "^6.38.0"
loguru = "^0.6.0"
datamodel-code-generator = "^0.11.19"
SQLAlchemy = "^1.4.35"
psycopg2-binary = "^2.9.3"
selenium = "^4.2.0"
webdriver-manager = "^3.7.0"
PyYAML = "^6.0"
pydantic-yaml = "^0.6.3"
fastapi = {extras = ["uvicorn"], version = "^0.78.0"}
uvicorn = "^0.17.6"
retry = "^0.9.2"
dependency-injector = "^4.39.1"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}
pylint = "^2.12.2"
matplotlib = "^3.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
known_first_party = ["nagasaki"]
profile="black"
line_length=88

[tool.black]
line_length=88

